requirements.txt
Flask==2.2.2
Flask-PyMongo==2.3.0
pymongo==4.3.3
Werkzeug==2.2.2
Jinja2==3.1.2
click==8.1.3
dnspython==2.2.1
python-dotenv==1.0.0
prometheus_client==0.17.1